Make floating window re-usable if already open
The issue could be solved by clearing the possibly former established
auto command to close the preview window. Clearing this first always
prevent any issues that the window disappears.

Furthermore the sub-module should become more independent, therefore
the main module makes the decision to add the auto command or not.